Extract this out, share it, hopefully covers travis
